refactor: move the `getYarnVersion` function to `src/yarn/version.ts`
