+++ src/yarn/index.ts
@@ -1,5 +1,6 @@
 import { exec, getExecOutput } from "@actions/exec";
 export { yarnInstall } from "./install.js";
+export { getYarnVersion } from "./version.js";
 
 export async function enableYarn(): Promise<void> {
   await exec("corepack", ["enable", "yarn"], { silent: true });
@@ -14,11 +15,4 @@ export async function getYarnConfig(name: string): Promise<string> {
     },
   );
   return JSON.parse(res.stdout).effective;
-}
-
-export async function getYarnVersion() {
-  const res = await getExecOutput("corepack", ["yarn", "--version"], {
-    silent: true,
-  });
-  return res.stdout.trim();
 }
