+++ src/yarn/version.test.ts
@@ -1,0 +1,29 @@
+import { jest } from "@jest/globals";
+
+jest.unstable_mockModule("@actions/exec", () => ({
+  getExecOutput: jest.fn(),
+}));
+
+it("should get Yarn version", async () => {
+  const { getExecOutput } = await import("@actions/exec");
+  const { getYarnVersion } = await import("./version.js");
+
+  jest.mocked(getExecOutput).mockReset().mockResolvedValueOnce({
+    exitCode: 0,
+    stdout: "1.2.3",
+    stderr: "",
+  });
+
+  const version = await getYarnVersion();
+
+  expect(getExecOutput).toHaveBeenCalledTimes(1);
+  expect(getExecOutput).toHaveBeenCalledWith(
+    "corepack",
+    ["yarn", "--version"],
+    {
+      silent: true,
+    },
+  );
+
+  expect(version).toEqual("1.2.3");
+});
