+++ src/yarn/version.ts
@@ -1,0 +1,8 @@
+import { getExecOutput } from "@actions/exec";
+
+export async function getYarnVersion() {
+  const res = await getExecOutput("corepack", ["yarn", "--version"], {
+    silent: true,
+  });
+  return res.stdout.trim();
+}
